@azalutsky 2023-01-17
Thanks guys, if anyone has any other tricks to speed up planning times- I'd love to know!