@azalutsky 2023-01-18
I am noticing however with multiple robots that there is a bug somewhere with these actions. 

Independent of the timeout, but the timeout alleviates it a little bit. 

For dual robot, the done flag claims it plans both and completes it however only the first robot in the group executes the trajectory while the second robot stays put. 

Seems to be a bug. I started tracking it but if someone has an idea I'd love to hear your thoughts as I investigate. 

I posted in discord's bug's channel about it. 

Would love a second opinion.