@rhaschke 2023-03-19
@ros-planning/moveit-maintainers, why does this code use busy waiting??? There are [condition variables](https://en.cppreference.com/w/cpp/thread/condition_variable) for this purpose...