@samialperen 2023-03-20
@azalutsky I did not see a performance increase even after this PR. I was expecting it would increase the performance but did not happen. Did you manage to find any other tricks to decrease planning time for humble?