@azalutsky 2023-03-20
If you use move group interface, when you call plan the action has this 100ms latency. 
What planning times are you seeing?