@azalutsky 2023-03-21
Is it possible you're not using moveit2 via source? Did you perhaps install moveit2 through apt-get or didn't colcon build with making sure that the package was built by source? I believe foxy requires some additional output.