@samialperen 2023-03-21
I built it from source and ros2 pkg prefix shows my workspace. For the reference, I am using moveit humble, (2.5.4)