@samialperen 2023-03-21
I find another issue stating that foxy has better performance than humble when it comes to planning times, but I could not test it on my system since I have too many dependencies for humble, I can not change it to foxy easily