@samialperen 2023-03-21
@azalutsky you were right. Somehow, there is a package collision even if ros2 pkg prefix shows the right package. After applying the fix manually, planning time reduced from 600 ms to 170ms.